+++ cookbooks/maxlab_dhcp/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 Deploy a DHCP server
 
+# 0.1.2
+
+* Using ::default instead of ::deploy across all cookbooks.
+
 # 0.1.1
 
 First initial working release